# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pokemon API",
        "description": "API for consuming Pokemon data",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local development server"
        }
    ],
    "paths": {
        "/api/pokemon": {
            "get": {
                "tags": [
                    "Pokemon"
                ],
                "summary": "Get pokemon list",
                "operationId": "d79c5d159e16c9963e038db7869a59d8",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "skip",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "bulbasaur"
                                            },
                                            "type": {
                                                "type": "string",
                                                "example": "grass, poison"
                                            },
                                            "image": {
                                                "type": "string",
                                                "example": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/1.png"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            }
        },
        "/api/pokemon/search": {
            "get": {
                "tags": [
                    "Pokemon"
                ],
                "summary": "Search pokemon by name",
                "operationId": "dd6996ecfd0830f65e5301864a91e4d0",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Pokemon name to search",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "pikachu"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "pikachu"
                                        },
                                        "type": {
                                            "type": "string",
                                            "example": "electric"
                                        },
                                        "image": {
                                            "type": "string",
                                            "example": "https://raw.githubusercontent.com/PokeAPI/sprites/master/sprites/pokemon/25.png"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pokemon not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No se pudo obtener información"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service unavailable"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Pokemon",
            "description": "Pokemon API"
        }
    ]
}